--- service/src/modules/eventManagement/endpoint/delete_object.go
@@ -1,10 +1,0 @@
-package endpoint
-
-import (
-	"context"
-)
-
-func (e *Endpoint) DeleteObject(ctx context.Context, id uint64) error {
-	err := e.usecase.DeleteObject(ctx, id)
-	return err
-}
