--- service/src/modules/eventManagement/transport/handler/create_object.go
@@ -1,29 +1,0 @@
-package handler
-
-import (
-	"encoding/json"
-	"net/http"
-
-	helperresponse "github.com/jabardigitalservice/super-app-services/event/src/helper/http/response"
-	"github.com/jabardigitalservice/super-app-services/event/src/modules/eventManagement/entity"
-	httpresponse "github.com/jabardigitalservice/super-app-services/event/src/response/http"
-)
-
-func (h *Handler) CreateObject(w http.ResponseWriter, r *http.Request) {
-	var obj entity.Object
-	if err := json.NewDecoder(r.Body).Decode(&obj); err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	// Create a context for the request
-	ctx := r.Context()
-
-	_, objData, _, _, err := h.endpoint.CreateObject(ctx, obj)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	helperresponse.Render(w, r, h.responseMapping, httpresponse.Created, objData, nil)
-}
