--- service/src/modules/eventManagement/usecase/delete_object.go
@@ -1,9 +1,0 @@
-package usecase
-
-import (
-	"context"
-)
-
-func (uc *Usecase) DeleteObject(ctx context.Context, objectID uint64) error {
-	return uc.repo.DeleteObject(ctx, objectID)
-}
