--- service/src/modules/eventManagement/usecase/get_object_by_id.go
@@ -1,16 +1,0 @@
-package usecase
-
-import (
-	"context"
-
-	"github.com/jabardigitalservice/super-app-services/event/src/modules/eventManagement/entity"
-)
-
-func (uc *Usecase) GetObjectByID(ctx context.Context, objectID uint64) (entity.Object, error) {
-	obj, err := uc.repo.GetObjectByID(ctx, objectID)
-	if err != nil {
-		return entity.Object{}, err
-	}
-
-	return obj, nil
-}
