--- service/src/modules/eventManagement/usecase/update_object.go
@@ -1,16 +1,0 @@
-package usecase
-
-import (
-	"context"
-
-	"github.com/jabardigitalservice/super-app-services/event/src/modules/eventManagement/entity"
-)
-
-func (uc *Usecase) UpdateObject(ctx context.Context, obj entity.Object) (entity.Object, error) {
-	updatedObject, err := uc.repo.UpdateObject(ctx, obj)
-	if err != nil {
-		return entity.Object{}, err
-	}
-
-	return updatedObject, nil
-}
