+++ service/src/modules/object/entity/entity.go
@@ -7,17 +7,6 @@ import (
 )
 
 type (
-	Organization struct {
-		Id          uint64    `db:"id"`
-		Name        string    `db:"name"`
-		Email       string    `db:"email"`
-		Address     string    `db:"address"`
-		PhoneNumber string    `db:"phone_number"`
-		Logo        string    `db:"logo"`
-		CreatedAt   time.Time `db:"created_at"`
-		UpdatedAt   time.Time `db:"updated_at"`
-	}
-
 	SocialMedia struct {
 		Name string `json:"name"`
 		Link string `json:"link"`
