+++ service/src/modules/object/transport/handler/handler.go
@@ -3,7 +3,7 @@ package handler
 import (
 	"github.com/jabardigitalservice/super-app-services/event/src/app"
 	helperresponse "github.com/jabardigitalservice/super-app-services/event/src/helper/http/response"
-	"github.com/jabardigitalservice/super-app-services/event/src/modules/eventManagement/endpoint"
+	"github.com/jabardigitalservice/super-app-services/event/src/modules/object/endpoint"
 	httpresponse "github.com/jabardigitalservice/super-app-services/event/src/response/http"
 )
