+++ service/src/services/main.go
@@ -58,7 +58,7 @@ func main() {
 	}))
 
 	router.Mount("/v1/ping", module.Ping.GetHttpRouter())
-	router.Mount("/v1/event", module.EventManagement.GetHttpRouter())
+	router.Mount("/v1/event", module.Object.GetHttpRouter())
 
 	var err = app.RunHttp()
 	if err != nil {
